int was used indisplay transition animation function.. corrected it..i used it before

+++ Bootloader/Periphery/ILI9341/ILI9341Device.cpp
@@ -423,7 +423,7 @@ void ILI9341Display::DisplayTransitionAnimation(bool& transitionActive)
         SendCommandPMP(ILI9341_RAMWR);
 
         uint16_t offset = ((float)(_transitionCounter) / (float)(255)) * ILI9341_TFTHEIGHT;
-        for(int index = 0; index < _framebufferSize; index++){
+        for(uint16_t index = 0; index < _framebufferSize; index++){
             if(index%ILI9341_TFTHEIGHT < offset){
                 WritePMP(_backFramebuffer[index]);
             }
